allow data transfer in batch
